+++ src/config/initializers/swagger.rb
@@ -1,3 +1,0 @@
-#GrapeSwaggerRails.options.url      = '/api/swagger_doc'
-GrapeSwaggerRails.options.url      = '/swagger_doc'
-GrapeSwaggerRails.options.app_url  = 'http://localhost:9292'