+++ src/config/routes.rb
@@ -1,5 +1,4 @@
 Rails.application.routes.draw do
   puts "Loading routes."
   mount API => '/'
-  mount GrapeSwaggerRails::Engine => '/'
 end