+++ CHANGELOG.md
@@ -1,1 +1,10 @@
-- 0.0.1 - Initial build
+
+0.0.3 
+=====
+* Removed grape-swagger-rails, in favor of DIY.
+* Added nginx
+* Added swagger-ui 2.1.5 M2
+* Hid swagger_doc from api exposure
+
+0.0.1 - Initial build
+=====================